Bagas

+++ app/Http/Controllers/admin/PengumumanController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
 use App\Models\Pengumuman;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 
 class PengumumanController extends Controller {
     public function index() {
@@ -21,22 +22,42 @@ class PengumumanController extends Controller {
         $simpan = DB::table('pengumumans')->insert([
             'judul'=>$request->post('judul'),
             'isi'=>$request->post('isi'),
+            'isi_preview'=> Str::substr($request->isi, 0, 50),
+            'slug'=> Str::slug($request->judul, '-')
         ]);
-
         return redirect()->route('admin.pengumuman');
     }
 
+    public function read($id) {
+        $pengumumans = Pengumuman::find($id);
+        return response()->json($pengumumans);
+    }
+
     public function delete($id) {
-        $pengumuman = Pengumuman::find($id);
-        $pengumuman->delete();
-        return redirect()->route('admin.pengumuman'); 
+        if ($id == null) {
+            return redirect()->route('admin.pengumuman');
+        }
+        try {
+            $pengumuman = Pengumuman::find($id);
+            $pengumuman->delete();
+            return redirect()->route('admin.pengumuman')->with('success', 'Data berhasil dihapus');
+        } catch (\Throwable $th) {
+            return redirect()->back()->with('error', 'Data gagal dihapus');
+        }
     }
 
-    public function edit() {
-        return view('admin.pengumuman.edit');
+    public function edit($id) {
+        $data['pengumumans'] = Pengumuman::where('id', $id)->first();
+        return view('admin.pengumuman.edit', $data);
     }
 
-    public function update() {
-
+    public function update(Request $request, $id) {
+        $pengumumans = Pengumuman::find($id);
+        $pengumumans->judul = $request->input('judul');
+        $pengumumans->isi = $request->input('isi');
+        $pengumumans->slug = Str::slug($pengumumans->judul, '-');
+        $pengumumans->isi_preview = Str::substr($pengumumans->isi, 0, 50);
+        $pengumumans->save();
+        return redirect()->route('admin.pengumuman');
     }
 }
